# Speedup for object tree crawling: fetch the whole tree and calculate