@mildbyte 2018-11-21
Coverage dropped because two object meta CRUD methods aren't now called by anything. I think we'll worry about it after the API refactor/reorganising.